refactor: controller 명, protocol 명 수정
