+++ JuiceMaker/JuiceMaker/Controller/FruitStockViewController.swift
@@ -8,7 +8,7 @@
 import UIKit
 
 
-class SecondViewController: UIViewController, FruitStoreViewController {
+class FruitStockViewController: UIViewController, StoreViewUpdatable {
     @IBOutlet weak var strawberryLabel: UILabel!
     @IBOutlet weak var bananaLabel: UILabel!
     @IBOutlet weak var pineappleLabel: UILabel!
@@ -20,7 +20,13 @@ class SecondViewController: UIViewController, FruitStoreViewController {
     
     override func viewDidLoad() {
         super.viewDidLoad()
-        labels = [strawberryLabel : .strawberry, bananaLabel : .banana, pineappleLabel : .pineapple, kiwiLabel : .kiwi, mangoLabel : .mango]
+        setLabels()
         configureFruitStoreUI()
+    }
+}
+
+extension FruitStockViewController {
+    func setLabels() {
+        labels = [strawberryLabel : .strawberry, bananaLabel : .banana, pineappleLabel : .pineapple, kiwiLabel : .kiwi, mangoLabel : .mango]
     }
 }
